Update Docker configuration and setup script
A new 'config.toml' file has been added to allow external access to the application's HTTP server. Updates to 'tsn.dockerfile' include copying the new 'config.toml' and setting up a new ENV variable 'PRIVATE_KEY'. Changes to 'setup.sh' script involve better process handling, error checks, and conditional config file creation. A new smoke test has also been added.

+++ truflation/scripts/setup.sh
@@ -6,16 +6,27 @@ set -e
 cd "$(dirname "$0")"
 
 # run it in parallel, assign process pid to kwild_pid
-../../.build/kwild --autogen & kwild_pid=$!
+../../.build/kwild --autogen &> /dev/null & kwild_pid=$!
+
+function cleanup {
+  echo "Killing kwild process $kwild_pid"
+  kill -9 $kwild_pid
+}
+
+# trap the cleanup function to SIGINT, SIGTERM and EXIT signals
+trap cleanup SIGINT SIGTERM EXIT
 
 # to make sure kwild is ready
 sleep 5;
 
-mkdir -p ~/.kwil_cli
-echo "{\"private_key\":\"$PRIVATE_KEY\",\"grpc_url\":\"http://localhost:8080\",\"chain_id\":\"\"}" > ~/.kwil_cli/config.json
+
+# if $PRIVATE_KEY is setup and config does not exist, we create with
+if [ -n "$PRIVATE_KEY" ] && [ ! -f ~/.kwil_cli/config.json ]; then
+  mkdir -p ~/.kwil_cli
+  echo "{\"private_key\":\"$PRIVATE_KEY\",\"grpc_url\":\"http://localhost:8080\",\"chain_id\":\"\"}" > ~/.kwil_cli/config.json
+fi
 
 # smoke test about kwil-cli
-echo "SMOKE TEST"
 test_content=$(../../.build/kwil-cli database list --self)
 
 # if contains Error, error out
@@ -38,5 +49,14 @@ python ./produce_source_maps/process_all.py
 ./database_deploy.sh --skip-drop
 ./database_add_primitives.sh
 
-echo "Killing kwild process $kwild_pid"
-kill -9 $kwild_pid
+sleep 10;
+
+# smoke test
+output=$(../../.build/kwil-cli database call -a=get_index date:"2023-01-01" date_to:"2023-12-31" -n=com_numbeo_us_bread_3m_avg)
+# output should contain `2023-12-01` to be sure it worked
+if [[ $output != *"2023-12-01"* ]]; then
+  echo "Smoke test failed: $output"
+  exit 1
+fi
+
+echo "Success!"
